requirements.txt
python-telegram-bot==13.15
transformers
torch
Flask
twilio
python-dotenv
